feat(accelerate): add China stops to event lineup
Share China roadshow data across the homepage and China page so all 2026
stops appear consistently, with reusable carousel controls and motion support.
