+++ apps/accelerate/src/components/homepage/index.ts
@@ -4,6 +4,7 @@ export { AuroraWave, HeroWaveWrapper } from "./AuroraWave";
 export { LiveBanner } from "./LiveBanner";
 export { EventLineup } from "./EventLineup";
 export { EventCard } from "./EventCard";
+export { CarouselArrow } from "./CarouselArrow";
 export { Highlights } from "./Highlights";
 export { VideoCarousel } from "./VideoCarousel";
 export { KeyStats } from "./KeyStats";
